Ref  Разбросал main.c по файлам для нормы
